Updated PostPolicy to handle authorization checks

+++ app/Http/Controllers/UserProfileController.php
@@ -5,17 +5,14 @@ namespace App\Http\Controllers;
 use App\Models\Post;
 use App\Models\User;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class UserProfileController extends Controller
 {
 
-    /**
-     * Display the specified resource.
-     */
     public function showUserPosts($id)
     {
-        
+
         $user = User::find($id);
 
         if (!$user) {
@@ -26,40 +23,40 @@ class UserProfileController extends Controller
             return redirect()->back()->with('error', 'No posts found for this user.');
         }
         $posts = $user->posts;
-       
+
         $postCount = $user->posts->count();
 
-        return view('profile.post.show', compact('posts','user', 'postCount'));
+        return view('profile.post.show', compact('posts', 'user', 'postCount'));
     }
 
     /**
      * Show the form for editing the specified resource.
      */
-    public function edit(Post $post)
+    public function editUserPosts(Post $post)
     {
-
-        return view('post.edit', compact('post'));
+        Gate::authorize('update', $post);
+        return view('profile.post.edit', compact('post'));
     }
 
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, Post $post)
+    public function updateUserPosts(Request $request, Post $post)
     {
-
+        Gate::authorize('update', $post);
         $post->update($request->all());
 
-        return redirect()->route('profile.show', $post->user_id)->with('success', 'Post updated successfully.');
+        return redirect()->route('dashboard')->with('success', 'Post updated successfully.');
     }
 
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Post $post)
+    public function deleteUserPosts(Post $post)
     {
-
+        Gate::authorize('delete', $post);
         $post->delete();
 
-        return redirect()->route('profile.show', $post->user_id)->with('success', 'Post deleted successfully.');
+        return redirect()->back()->with('success', 'Post deleted successfully.');
     }
 }
